prov/rxm: fix address comparison to remove duplicate connections
The address comparison to determine whether to reject or keep a duplicate
connection was incorrect and falsely identifying too many loopback
connections. The address to compare against should be the local EP's
address, not the peer's address.

This fix exposes an issue with the reject/close path which was not being
taken before. A race condition during simultaneous connections was
occurring where both sides sent a connection request. One side received
a connection request, accepted, and closed its outgoing connection.
However, the connection request had already been received by the peer,
causing the peer to incorrectly identify that connection as a lost
connection, causing send failures for any outstanding sends on that
active connection.

This patch could potentially cause issues in how lost connections are
managed, but testing to date has not shown any issues... yet...

Signed-off-by: aingerson <alexia.ingerson@intel.com>