prov/efa: adjust names of some functions and variables
This patch make the following variable name changes:

In struct rxr_ep,
	posted_bufs_shm     -> shm_rx_pkts_posted,
	rx_bufs_shm_to_post -> shm_rx_pkts_to_post,
	posted_bufs_efa     -> efa_rx_pkts_posted,
	rx_bufs_efa_to_post -> efa_rx_pkts_to_post.

It also make the following function name changes:

	rxr_ep_post_user_buf -> rxr_ep_post_user_recv_buf
        rxr_ep_post_prov_buf -> rxr_ep_post_internal_rx_pkt
        rxr_ep_bulk_post_prov_buf -> rxr_ep_bulk_post_internal_rx_pkts
        rxr_ep_progress_post_prov_buf -> rxr_ep_progress_post_internal_rx_pkts

The new names are more accurate in describe what the functions
are doing: which is to post a RX packet to device to receive data.

The function name rxr_ep_post_user_recv_buf() is different because it
converts an user buff into a RX packet, and post it to the device.

Signed-off-by: Wei Zhang <wzam@amazon.com>