fabtests/ubertest: Add CQ_FORMAT to ubertest
The default CQ_FORMAT tested in ubertest was CQ_FORMAT_DATA. This changeset adds
testing for all the CQ_FORMATS.

This also adds verification for valid fields including op_context, data, tag, length
and flags for UNIT testing. IF this is not set in the config file it defaults to
appropriate formats.

It skips invalid combinations such as not using the tag format for TAGGED tests or
using TAGGED format for non-tag tests which fails for providers that do not support
FI_TAGGED (verbs).

This adds the check_tag and check_ctx in ft_xcontrol to match the receive tag and
context to the sent ones.

This adds the testing to tcp config file.

Signed-off-by: Juee Himalbhai Desai <juee.himalbhai.desai@intel.com>