fabtests: add missing ft_init calls for hmem init
A couple of fabtests are not calling ft_init, specifically the atomic
tests and rdm multi client test. Those tests use the general memory
allocation path and argument parsing so it is possible for them to call
hmem alloc without calling hmem init. Add calls to ft_init so hmem is
initialized.

Signed-off-by: Robert Wespetal <wesper@amazon.com>