prov/efa: Fix the mishandling of prefix size when receiving data
Currently, the msg prefix size is not included in the offsets
of data copies on the receiver side, which causes data corruptions when
FI_MSG_PREFIX is specified in hint->mode. This patch expands the truncation
size of rx / read entries to include the prefix size, and shift the
offset of local iov accordingly.

Signed-off-by: Shi Jin <sjina@amazon.com>