prov/efa: Remove redundant variable "lower_efa_prov".
Since the efa_fabric and rxr_fabric functionality was squashed a while back,
this variable was redundant and added to unnecessary code complexity.

This change simplifies the code by removing this variable and replacing
with existing functionality as needed. Going forward this will make the code
easier to understand and maintain.

Signed-off-by: Rich Welch <rlwelch@amazon.com>