prov/efa: use a static connid for DGRAM endpoint
The connid was introduced to EFA address to distinguish two
endpoints with same GID+QPN.

This distinction is important for RDM endpoint because packets
in RDM communcation protocol are not independent, thus it is
necessary to process a RDM packet according to its connid.

However, this distinction is not important for DGRAM endpoint
because each DGRAM packet is independent.

Therefore, this patch used a static connid (0x0)
for any dgram endpoint.

This patch also replaced the checking/handling of existing reverse AV
entry in function efa_conn_alloc with an assertion. This is because
with a static connid, efa_conn_alloc will not be called if there is
an existing entry in reverse AV with same GID+QPN.

Signed-off-by: Wei Zhang <wzam@amazon.com>