prov/efa: Adding some checking around enabled shared memory transfers
There are currently a couple of places where a global variable is used
to control whether or not shared memory is used for intranode
communnication. This can be dangerous if for some reason this gets
out of sync with the actual underlying structures.

This patch changes things to that we primarily check the underlying
structures, and only set those after checking the global variable
once. There are some occurances where making this change isn't
possible but for the most part this patch makese these checks more
defensive.

Also, in some places the check on the global variable is redundant
and any inconsistency would represent an error state, so these checks
were changes to asserts.

Signed-off-by: Rich Welch <rlwelch@amazon.com>