prov/tcp: Fix possible memory leak of cm_ctx
If an endpoint is closed during the middle of trying to
establish a connection, the associated cm_ctx may be left
dangling.  This happens when rxm identifies peer connections
and aborts one side.

To resolve, maintain a reference to the cm_ctx from the ep
while the cm_ctx is valid.  When closing the ep, free the
cm_ctx if it is still valid.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>