prov/efa: write rx completion upon send completion of RECEIPT
When delivery complete is requested, the sender will wait for
a RECEIPT packet from receiver.

Currently, on receiver side, RX completion is written when
data has been copied to application's receive buffer, before
RECEIPT's send completion. As a result, it may happen that
application tear down the endpoint upon receiving RX completion,
causing the RECEIPT packet not being sent, which will cause
the sender to wait for RECEIPT indefinitely.

This patch address the issue by delaying the writing
of RX completion to after getting the send completion of the
RECEIPT packet.

With this change, the function rxr_cq_handle_rx_completion()
has to be changed to handle the RX completion of a DC write
atomic.

Signed-off-by: Wei Zhang <wzam@amazon.com>