prov/verbs,rxm: Fix hook provider handling in the flow control ops
The credit based flow control mechansim uses functions exported via
the fi_open_ops() interface to coordinate between the rxm and verbs
layers. There are two issues when a hook provider sits between rxm
and verbs:

(1) These exported functions expect verbs objects (ep, domain, etc),
but here the hook objects are passed in.

(2) The 'rxm_send_credits' call expects the ep passed in has 'rxm_conn'
as the fid context, but here the fid context contains the hook ep, and
'rxm_conn' resides in the fid context of the hook ep.

To fix these, above functions need to check if the objects are from
the hook provider and adjust the behavior accordingly. In addition, a
utility function 'is_hook_fid' is added to help the checking.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>