core/collectives: Define memory registration for collectives
Introduce a new mr_mode bit, FI_MR_COLLECTIVE, to indicate if
a data buffer must be registered prior to being used as part of
a collective operation.  Enhance the mr registration flags to
cover collective operations.  This coverage was missing from the
original submission for collectives.  Buffers used with collectives
are now covered by the FI_SEND and FI_RECV flags.

Add a new FI_COLLECTIVE flag to registration to explicitly
indicate that a buffer will be used as part of a collective.  This
allows the provider to setup permissions based on the underlying
implementation, such as adding RMA access to such buffers.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>