prov/efa: remove usage of SMR_NAME_MAX
Currently, the function rxr_ep_addr_to_shm_name() and its
caller uses SHM_NAME_MAX to define the buffer size for the
shm endpoint name, which introduced a dependency on ofi_shm.h

This dependency is unnecessary because the length of shm endpoint
name can be derived from the name format EFA used.

This patch introduce a macro EFA_SHM_NAME_MAX, and used it
to replace the usage of SMR_NAME_MAX. It also removed the inclusion
of ofi_shm.h to remove the unnecessary dependency.

Signed-off-by: Wei Zhang <wzam@amazon.com>