prov/util: Add logic to detect previous memory patching
Currently, we cannot detect if another application patches memory relating to how
we monitor memory registration. This change adds this functionality so that if
another application has patched the memory in the same way we'll detect it before
we patch it (again).

If this happens, the ofi_intercept_symbol() function will return FI_EALREADY
under this condition.

Signed-off-by: Rich Welch <rlwelch@amazon.com>