prov/verbs: Avoid possible use-after-free of the rdma cm id
If an user makes the following calling sequence, it can
result in accessing the rdma cm id after it has been
destroyed:

receive CM connection request (info->handle)
fi_endpoint(... info->handle, &ep);
fi_close(ep);
fi_reject(info->handle);

The issue is that the info->handle is moved/owned by the
ep once it is created.  When the ep is closed, the underlying
rdma cm id is destroyed.  fi_reject then tries to reference that
same id.

To fix this, clear the rdma cm id from the connreq object.
When the ep is closed, it will destroy the id and associated
resources.  Add a check to reject to verify that the id is
still valid before accessing it.

While this fix isn't amazing, it will prevent crashes in the
above calling sequence without adding more complex state
tracking or racy code handling ep close with reject.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>