prov/rxm: Correcting the setting of conn_id in rx pkt control header
In the rxm_fake_rx_hdr function the rx_buf->pkt.ctrl_hdr.conn_id was
being incorrectly set to conn->remote_index.Since this interface is used
by the receiving side to support tcp optimizations for small messages
this value should be updated with the peer->index. The peer->index
is used to fetch the appropriate connection and using the remote index
causes the connection lookup to fail sometimes.

This patch provides fix for PR 6682.

Signed-off-by: Nikhil Nanal <nikhil.nanal@intel.com>