prov/efa: allow EFA SHM RX pkt pool to grow for unexp msgs
Currently the SHM RX packet pool is a fixed size based on the RX queue
size of SHM (1k). This can be exhausted by unexpected messages and cause
applications to hang when the RX queue is exhausted.

We had a similar bug with the EFA queues and added a copy out mechanism
to keep the RX queues hydrated. SHM is a bit different in that we do not
need to register the RX buffers, so this patch instead allows the SHM RX
packet pool to grow, and changes the type of the packet. Unexpected
messages will stay in the same buffer until processed, but the SHM RX
pool can now grow to keep 1k buffers posted with SHM. This will lead to
increased memory usage when unexpected messages are received over SHM,
but will avoid an extra memcpy.

Signed-off-by: Robert Wespetal <wesper@amazon.com>