prov/rxm: Add AV lock around modifications to the peer_addr
rxm_conn can reference a rxm_peer_addr.  In most cases, the conn
will only read data from the peer_addr, which is safe without
additional locks.  (Once the peer is created, it's address data
is static).  However, we need to hold the AV lock if the peer
data is modified.  This is missing in rxm_alloc_conn, where
the peer->refcnt is incremented.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>