prov/efa: Fix a bug in rxr_pkt_handle_send_error
When RNR happens, if the packet is associated with a send
operation, it's supposed to do one of the following two:
(1) write an error to the application, and release the packet
if the resource management is disabled;
(2) queue the packet, and send later.

We were missing a 'else', so we ended up doing both (1) and (2).

This patch fixes the bug and also does code refactoring to
make the logic easier to understand.

Signed-off-by: Jie Zhang <zhngaj@amazon.com>