prov/shm: return correct error code if peer is not initialized
The shm code assumes FI_ENOENT as the error code if the peer is not
initialized. The old code was using FI_ENOENT if the shm region did
not exist but then FI_EAGAIN if the shm existed but the peer has not
finish initializing, causing interpretation errors in the fi_av_insert
path.

Signed-off-by: aingerson <alexia.ingerson@intel.com>