prov/shm: fix possible sigbus error
If the client tries to map a peer between when the shm
was created and when it was truncated to the possible size,
the client might throw a SIGBUS error trying to access the
peer region's memory before it actually exists.
This adds a check to find the actual size of the shm file
and make sure that all smr_region fields are accessible before
we check them later.

Signed-off-by: aingerson <alexia.ingerson@intel.com>