prov/rxm: Do not ignore user specified default rx/tx sizes
Users may set the default rx/tx sizes through an environment
variable.  However, in rxm_info_to_rxm, the returned sizes
are forced set to a default value of 2k, if that is smaller.
Use the sizes specified by the user, or that returned by the
msg provider if smaller.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>