prov/rxm: Add call to retrieve number of known peers
There's a minor (i.e. should be harmless) race in rxm_freeall_conns.
The for-loop which checks for connections reads the maximum number
of peers in the AV (peer_pool->entry_cnt).  It's possible for this value
to change during the iteration of the loop.  Although an
existing peer cannot be removed while we have an active connection,
new entries could be added.

Update the code to remove the race, so that all accesses to the
AV peer_pool are serialized.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>