efa: Update mem monitor to default monitor
Revamp how EFA initializes the SYSTEM monitor to handle more
use cases and to use the default_monitor whenever possible.
This patch includes a check to verify that the memhooks monitor
can actually patch the binary (and someone else hasn't already
patched the binary) and will fall back to UFFD in the case that
we would use memhooks but the binary has been patched.  With
the switch to the default_monitor, EFA will also use any
import monitors that were set.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>