prov/efa: check if hmem device is initialized, add check to mr regattr
Instead of just relying on the sysfs file, validate that hmem support
was initialized for that particular device during core initialization,
and don't report FI_HMEM support if support is not available.

Also add some checks fo the mr regattr paths to validate whether support
exists. This isn't strictly necessary, but may help users when the
provider supports multiple devices. The current FI_HMEM flag doesn't
provide information on which devices are supported by the provider.

This doesn't contain changes around whether p2p support is available, and
falling back to copies when it is not, so p2p is still required for EFA +
FI_HMEM at this point. We'll also want to store information about p2p support
in efa_ep, and check that information in the registration path.

Signed-off-by: Robert Wespetal <wesper@amazon.com>