prov/cxi: Support fi_cntr_read/err with dev mem
When the counter writeback buffer is in device memory, the CXI provider
cannot directly reference the memory. It has to use bounce buffers to
read/write the writeback buffer.

Update the CXI provider to memcpy to/from device memory when it has to
write/read from the counter writeback buffer in GPU memory.

NETCASSINI-2864

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>