prov/cxi: Update configurable run-time environment options
Entering flow-control is no longer experimental and
always enabled; remove code and documentation to disable it.

On-loading of the rendezvous protocol requires message
matching be off-loaded. If rendezvous is on-loaded,
then force message offloading to be true, disabling
software managed EP.

Allow setting of RX matching mode to support setting of
"hardware" only, "software" only, and a "hybrid" that
starts with hardware matching, but when resources are
low will transition to hybrid mode, sharing matching
responsibility with software.

NETCASSINI-2412

Signed-off-by: Steve Welch <welch@hpe.com>