fabtests/ubertest: Fix bug for not checking verification count
The verification count tracks how many valid cq completions we got.
The verification only happens if comp_cq is set and not when comp_cntr
is set. Ubertest was not checking the verification count for all valid
cases and sometimes was reporting failures when using counters because
the verification is skipped for counters.

Signed-off-by: Juee Himalbhai Desai <juee.himalbhai.desai@intel.com>