prov/efa: Use FI_OPT_FI_HMEM_P2P option
Add state to track three things: 1) if libfabric core was able to initialize
devices we support, 2) whether we support p2p transfer for hmem, and 3) what
the user specified for p2p transfer requirements.

Add checks to validate this state before a transfer.

Finally, add this option to getopt/setopt handlers. We also add an EFA specific
environment variable to allow Open MPI to set this in a way that can be
overridden later, once it adds an Open MPI parameter for users to toggle this.

Signed-off-by: Robert Wespetal <wesper@amazon.com>