prov/tcp: Fix reporting RMA write CQ data
When receiving an RMA write with CQ data, we need to
add the FI_REMOTE_CQ_DATA flag to the completion to
indicate that the CQ data is valid.  We should also
report the CQ data.

This change updates setting the FI_REMOTE_CQ_DATA flag
on the CQ entry for the RMA write and adds a check to
report that data.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>