prov/tcp: Move internal xfer flags to own ctrl_flag field
The tcpx_xfer_entry flags field is used for writing the
CQ entry flags and for controlling data transfers.  The
control flags are marked off using an internal mask.
Add a new flag field to the xfer_entry and move the
internal flags to the new field.  This improves code
readability and eliminates the need to mask off the flags
when writing CQ entries.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>