fabtests/multinode_coll: Add test with start_addr != 0 and stride != 1
Add a test case for when av_set_attr.start_addr != 0 and
av_set_attr.stride != 1.

It runs an extra test with sum_all_reduce_test_run targetting half of
the ranks.

The test is now using a data of rank + base_data_value to be able to
check if rank 0 takes part or not in the collective.

Signed-off-by: Olivier Serres <oserres@google.com>