prov/tcp: Mask off non-CQ flags when reporting completions
The xfer_entry flags is set to a copy of the operational flags
specified as part of the operation.  However, the flags are only
used later to fill in the CQ entry.  Only store CQ entry related
flags and mask off other operational flags, the exception being
FI_COMPLETION, which is used to control whether a completion is
written or not.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>