core: Make output for printing fi_info fields consistent
The fi_info field name is 'nic', not 'nic_fid' or 'fid_nic'.
The string is inconsistent and should match the name of the
field itself.

Update reporting bus_type field from bus_attr to match the
field name, dropping the fi_ prefix.

Add missing tclass field to the output.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>