prov/efa: Do not claim support of 128 bit atomic ints
Due to the way memory is currently aligned when performing atomic operations,
the EFA provider was segfaulting when atomic operations were performed on 128
bit data types. This is because the compiler optimizes for instructions that
require 16 byte alignment and since we do not (currently) have that, the code
segfaulted.

Eventually we'll need to add support for these types, but for now this change
will explicitly disable support for these types in order to prevent a segfault.

Signed-off-by: Rich Welch <rlwelch@amazon.com>