prov/tcp: Allow binding a CQ with an SRX
In order to support trecv with the FI_PEEK flag, we need the
ability to generate a completion that's associated with a
shared receive context, and not necessarily a connected
endpoint.  Binding a CQ with the SRX allows us to direct
the completion to the CQ.

This also allows us to flush queued receive operations posted
to the SRX when its being destroyed.

This is a non-standard feature of the API that rxm will make
use of, but is a candidate for adding to the API.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>