prov/cxi: Avoid pointer queries if possible
Pointer queries cause an additional .2+ usecs of latency. This should be
avoided at all costs.

If FI_HMEM support was requested, cxip_map() would always do a pointer.
But if only AMD GPU (ROCR) device memory is supported and the default
ROCR memory monitor is the libfabric core ROCR memory monitor, an MR
cache find can be issued to try and find a cached MR before any pointer
query is issued. If a cached entry is not found, then a pointer query
and an MR cache search is issued which will allocate the MR cache entry.

This approach is ONLY valid if FI_HMEM_SYSTEM and FI_HMEM_ROCR is
supported due to how the libfabric core ROCR memory monitor is
implemented. Unlike the FI_HMEM_CUDA and FI_HMEM_ZE memory monitors, the
ROCR memory monitor relies on the HSA runtime to notify libfabric when a
ROCR buffer is freed. Thus, the ROCR memory monitor valid() operation is
a no-op (i/e is always true).

FI_HMEM_CUDA and FI_HMEM_ZE both implement the memory monitor valid()
operation which is used to detect freed buffers. The only place valid()
is called is when ofi_mr_cache_search() is called and
ofi_mr_cache_search() requires the iface to be set correctly. This
requires a pointer query.

This PR reduces 256-byte OSU latency from 3.12 to 2.85 usec.

NETCASSINI-3256

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>