prov/shm: distinguish between max shm name and max path name
The current implementation defines SMR_NAME_MAX as 256 as that aligns with the
limit for file names. However, this doesn't take into account the assumed path
prefix for shared memory files ("/dev/shm/"). Define SMR_PATH_MAX to account for
the prefix and modify smr_map_to_region to use SMR_PATH_MAX (the whole length)
since it is used with the path prefix.

Signed-off-by: aingerson <alexia.ingerson@intel.com>