prov/efa: skip registering MR with device for hmem buffer if no p2p
If p2p is not supported by the driver/device, don't bother registering the MR
with the device.

XXX: this is half baked/still needs some more thought on two things 1) we
depend on the device to provide registration keys so one sided cuda ops will
fail with this patch, and 2) need to think about shm registration too (for now
skip the shm registration as well since we turn off shm when fi_hmem is
enabled).

Signed-off-by: Robert Wespetal <wesper@amazon.com>