fabtests/ubertest: Add GPU testing and validation support
This commit adds GPU testing and validation support in ubertest
along with memory allocation and registration. The definition
of ft_reg_mr() has been changed to add the fi_info struct so that
the correct mr mode gets passed in.

This commit also includes a small bug fix for an  ubertest bug
where it wasn't handling FI_MR_PROV KEY mode correctly. That is
solved by passing a unique key into the mr registration.

All the ATOMICS test cases are skipped for now because ubertest
does not support ATOMICS testing for device memory.

Signed-off-by: Juee Himalbhai Desai <juee.himalbhai.desai@intel.com>