hmem/cuda: fix gdrcopy_dev_register()'s calculation of memory region.
gdrcopy_pin_buffer requires the input to be aligned on page boundary.

This patch fixed a bug in gdrcopy_dev_register() in calculation
of page aligned addresses:

if the memory region passed from application is exactly on page bounary,
this function will added an extra page size to the ending address.

Signed-off-by: Wei Zhang <wzam@amazon.com>