fabtests/benchmarks: Fix inject size setting
The '-j' option of the benchmark tests allows the user to specify the upper
limit of inject size. For example, one can disable inject calls by setting
the inject size to 0. This value is passed via the hints to fi_getinfo().
However, some providers (e.g. rxm) may report higher inject size than asked
in the hints. Take the value from the hints when '-j' option is used to honor
the user input.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>