fabtests/common: add key exchange support for device memory
The HMEM key exchange support is broken because it accesses the tx and rx buf directly
which could be device memory. To fix this, use a temporary buffer to build the rma iovs
and then copy that temp into and out of device memory using the ft_hmem_copy functions.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>