fabtests/common: remove need for OOB address exchange with hmem
Use a temporary buffer for calling fi_getname so we ensure a host
buffer is passed into the call (which does not support device memory).

Copy those contents into the tx buffer using the appropriate hmem copy
function.

This removes the need to force an OOB address exchange when using
device memory, aligning with what is needed for the RMA key exchange.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>