fabtests/rdm_rma_trigger: remove hardcoded tx/rx seq values
The rdm_rma_trigger test had hard coded values (4) which got
misaligned when an extra send/recv were added in the initialization path.

This removes the need for the hardcoded values by storing the start
tx and rx values after initialization so the test always expects the
correct number of completions during the test.

Signed-off-by: aingerson <alexia.ingerson@intel.com>