fabtests/recv_cancel: Remove requirement for unexpected msg handling
The fi_recv_cancel test assumes that the receive will process and
queue an unexpected message without a matching buffer present,
followed by processing an expected message.  This places a buffering
requirement on the provider which may not be true.

Fix this by reversing the order that the messages are posted to
remove the need for unexpected message queuing.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>