prov/rxm: Shutdown existing connections if we receive a new request
If we receive a connection request from the same peer address
and we're in the CONNECTED or ACCEPTING state, the new request
likely indicates that the old connection is no longer valid.
(We expect the lower layer to filter out true duplicate
connection requests).  When this occurs, close the older
connection and accept the new request.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>