prov/efa: handle queued local read request correctly
Progress engine will submit queued read request.

When doing that, it needs to retrieve the peer of the read
to check whether peer is in RNR backoff state.

Local read request (which is used to copy data from host
memory to device memory) does not have a peer, thus
the returned peer pointer would be NULL, therefore
a queued local read requst will cause segfault.

This patch handles that case correctly by not checking
peer state if peer is NULL, it also removed the assertion
of peer not being NULL.

Signed-off-by: Wei Zhang <wzam@amazon.com>