core/getinfo: Fix check if provider requires use of shared contexts
The tcp provider requires the use of shared contexts in order to
support the FI_TAGGED capability.  By default, if an application
does not specify a tx or rx cxt_cnt, the returned fi_info should
be 1.  Update the checks in ofi_check_ep_attr to see if the user
has specified a tx or rx cxt_cnt.  If no count was specified and
the provider's fi_info structure requires FI_SHARED_CONTEXT, fail
the info match.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>