core: Define XPU triggering capability
Many HPC and AI applications use GPUs or other accelerator devices
to perform computation work.  FI_HMEM support allows a NIC
provider to send and receive data stored on a GPU.  However,
the data transfer itself must still be initiated by the host
CPU.  This introduces delays from the time the data is ready to
be transfered until the network transfer is initiated.  Once
the GPU is done processing the data, it must signal the CPU
that the processing is complete, and then the CPU can request
the network transfer.

This patch introduces a new type of triggered operation, called
an XPU trigger.  An XPU trigger works as follows:

1. An endpoint is created with XPU triggering support.  By
default, all data transfers posted to the endpoint follow
the XPU triggering model.

2. The host CPU makes a data transfer call (e.g. fi_write)
with the FI_TRIGGER flag set.  The triggering event is
FI_TRIGGER_XPU.

3. A provider responds to the data transfer call by queuing
the transfer, but will not initiate the transfer until a
provider specified signal is set.  The signal may be signaled
by a device, such as a GPU or FPGA, referred to as an XPU.

4. Once an XPU writes the signal, the provider will initiate
the data transfer.

Details and restrictions are defined as part of the man page
updates.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>