src/hmem: pass through device and flags when querying memory interface
The ofi_get_hmem_iface calls iface.is_addr_valid which takes optional device and
flags fields that return the device and flags information for the memory.

Add these fields to ofi_get_hmem_iface to pass through to the is_addr_valid call
in case the caller needs the device and flags information. If the caller does not
need the information, NULL can be provided to skip the extra look up in is_addr_valid.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>