prov/hook_hmem: add HMEM memory registration hook
The ofi_hook_hmem provider can be enabled to help applications register HMEM memory
if the application is not fully GPU aware. The hook will enable the MR support
by hooking into all transfer calls to query the memory and register it with
the core provider with appropriate iface, device, and flags as necessary, keeping track
of reference counters as well.

The hooking provider will also track all outstanding transfers in order to
cache the memory properly and register/close MRs when they are no longer being
used.

In order to track transfers properly, the hooking provider cannot enable
selective completions in the core provider but hides this limitation from the
application by storing the CQ format and determining whether the application
expects a completion from each transfer, writing one to the app's CQ when
necessary.

This is step 1 in the hooking provider implemention. Step 2 involves refactoring
the generic hooking provider implementation to hook into the getinfo call as well.
Currently, the hooking providers are defined by whether or not the have a getinfo call.
For fully correct implementation, the hmem hook needs to intercept the getinfo call
in order to add FI_HMEM and FI_MR_HMEM to query for HMEM support. Otherwise,
the application needs to add FI_HMEM and FI_MR_HMEM, essentially falsely advertising
FI_MR_HMEM support in their application.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>