prov/cxi: Support FI_HMEM_ZE
FI_HMEM_ZE uses the cxi-core DMA buf support. If memory is identified as
FI_HMEM_ZE, the corresponding DMA buf FD and address range is query. The
DMA buf FD is needed to map the Intel GPU BAR with the CXI NTA tables
and the entire GPU buffer address range is needed to determine the
proper DMA buffer offset to be applied.

NETCASSINI-2047

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>
[welch@hpe.com: minor cleanup, add build prep for ze]
Signed-off-by: Steve Welch <welch@hpe.com>