prov/util: Restructure collective av set creation/destruction
There is a memory leak of the coll_mc when the parent AV is
destroyed.  To fix the leak, we simplify and restructure the
creation of AV sets.  This may seem like overkill, but is
necessary to fixup object reference counting.

The current code associates a base collective multicast group
(coll_mc) with an AV the first time an AV set is created.  This adds
a coll_mc to the AV, which is what gets leaked.  However, that
coll_mc itself creates an av_set.  Part of the generic creation
of an av_set, a coll_mc is created.  The result is that we have
AV -> coll_mc -> av_set -> coll_mc.  So, there's a double coll_mc
created for the AV.

We can simplify the code by having the AV create an initial av_set
when needed, and let create av_set create the coll_mc.  Then use
that coll_mc and avoid the duplicate coll_mc creation.  This
keeps the av_set to coll_mc objects the same regardless of
whether we allocate them internally or in response to the user
calling fi_av_set().

Finally, to fix the memory leak, when destroying the AV, we just
need to call fi_close() on the av_set.  The resulting code has
common paths for av_set and coll_mc creation/initialization/
destruction.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>